Initial Commit
Created a javascript tool which uses html canvas to render a user defined string onto frames of a lakitu "REVERSE" sprite from mk64 and render those frames out to an animated gif.

Constructed an html page frontend for the tool.

Created a custom font to try to match the font seen on the original sprite.